minimize `String` copies and allocations in `choices`.
We used to allocate and construct a new `String` for every single entry in the choose and order lists. This copying was unnecessary since we never modified the `String`s in place, so I rewrote everything to take and return `&str` where possible.

Because we allow choosing from a numerical range, it's not possible to return `&str` from `do_choose`. However, this isn't a huge deal since `say` takes `Into<String>` so at least one allocation has to happen each time these commands are run.